+++ modules/apps/headless/headless-delivery/headless-delivery-api/src/main/resources/com/liferay/headless/delivery/dynamic/data/mapping/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
